RFIDNet/requirements.txt
adapter==0.1
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Django==4.2.7
django-leaflet==0.29.0
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
Pillow==10.1.0
psycopg==3.1.14
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==3.0.1
